prediction_service/requirements.txt
torch
scikit-learn
pandas
numpy
joblib
fastapi
uvicorn
sqlalchemy
psycopg2-binary
asyncpg
pydantic_settings